# specs/001-password-manager/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Password Manager API
  version: 1.0.0
  description: >
    Zero-knowledge password manager. The server stores and authorizes ciphertext it cannot read.
    Every field with `format: envelope` is client-encrypted per contracts/crypto-envelope.md.
    See contracts/README.md for the error model, authorization matrix, and known limitations.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
  /auth/kdf-params:
    post:
      tags: [auth]
      summary: Fetch KDF parameters for an email (pre-login)
      description: >
        Returns parameters for ANY email, real or not, so the endpoint cannot be used to enumerate
        accounts (FR-003). For unknown emails, parameters are derived deterministically from the
        email so they are stable across calls.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200':
          description: KDF parameters
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KdfParams' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/register:
    post:
      tags: [auth]
      summary: Create an account and its personal vault
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                [email, authHash, kdfParams, wrappedUserKey, publicKey, wrappedPrivateKey,
                 wrappedPersonalVaultKey, personalVaultName, recoveryAcknowledged]
              properties:
                email: { type: string, format: email }
                authHash:
                  type: string
                  description: Argon2id(MasterKey, salt=masterPassword). NEVER the master password itself.
                kdfParams: { $ref: '#/components/schemas/KdfParams' }
                wrappedUserKey: { type: string, format: envelope }
                publicKey: { type: string, description: RSA-4096 SPKI, base64url, not encrypted }
                wrappedPrivateKey: { type: string, format: envelope }
                wrappedPersonalVaultKey: { type: string, format: envelope }
                personalVaultName: { type: string, format: envelope }
                recoveryAcknowledged:
                  type: boolean
                  description: >
                    MUST be true. The user has been told a forgotten master password means permanent,
                    irreversible loss of all vault contents (FR-010). Registration is refused otherwise.
      responses:
        '201': { description: Account created; verification email sent }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { description: Email already registered }
  /auth/login:
    post:
      tags: [auth]
      summary: Authenticate with the derived auth hash
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, authHash]
              properties:
                email: { type: string, format: email }
                authHash: { type: string }
                deviceLabel: { type: string }
      responses:
        '200':
          description: >
            Session established. If TOTP is enrolled, `status` is `totp_pending` and the session may
            only reach /auth/totp/* until upgraded (see README limitation 1).
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, enum: [active, totp_pending] }
                  keyring: { $ref: '#/components/schemas/Keyring' }
                  newDevice:
                    type: boolean
                    description: True if unrecognized; triggers a security notification email.
        '401':
          description: >
            AUTH_FAILED — returned identically for unknown account and wrong password, with a
            constant-time floor (FR-003).
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /auth/master-password:
    put:
      tags: [auth]
      summary: Change the master password
      description: >
        Requires the CURRENT master password, not merely a valid session (research.md §9) — a stolen
        session must not be sufficient to cause permanent lockout. Rewraps the UserKey only; no
        secret is re-encrypted (FR-005).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentAuthHash, newAuthHash, newWrappedUserKey, newKdfParams]
              properties:
                currentAuthHash: { type: string }
                newAuthHash: { type: string }
                newWrappedUserKey: { type: string, format: envelope }
                newKdfParams: { $ref: '#/components/schemas/KdfParams' }
      responses:
        '204':
          description: >
            Changed. Every other session is stamped reauthRequired and will be refused with
            401 REAUTH_REQUIRED until it calls POST /auth/reauth (FR-072, FR-073). Sessions are not
            terminated. A notification email is sent.
        '401': { description: Current master password incorrect }
  /auth/reauth:
    post:
      tags: [auth]
      summary: Clear a session's re-authentication requirement after a master password change
      description: >
        The only route a session stamped reauthRequired may call (FR-073). Enforcement is
        server-side: a client that already unwrapped the UserKey can ignore any client-side prompt,
        so the session guard refuses every other route regardless of what the client does
        (research.md §13). The second factor is NOT re-challenged (FR-074).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [authHash]
              properties:
                authHash:
                  type: string
                  description: AuthHash derived from the NEW master password.
      responses:
        '200':
          description: Flag cleared; the rewrapped keyring is returned so the client can rebuild its key chain.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Keyring' }
        '401': { description: AUTH_FAILED — the AuthHash does not match the new master password }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: pm_session
...
  responses:
    BadRequest:
      description: VALIDATION_FAILED or ENVELOPE_MALFORMED
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: RATE_LIMITED
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object }
    KdfParams:
      type: object
      properties:
        algorithm: { type: string, enum: [argon2id, pbkdf2] }
        memoryKib: { type: integer, example: 65536 }
        iterations: { type: integer, example: 3 }
        parallelism: { type: integer, example: 1 }
    Keyring:
      type: object
      description: Wrapped key material. Useless without the master password.
      properties:
        wrappedUserKey: { type: string, format: envelope }
        wrappedPrivateKey: { type: string, format: envelope }
        publicKey: { type: string }
...